refactor: extract run metrics builder module
